+++ src/components/menu-item/menu-item.styles.ts
@@ -147,4 +147,9 @@ export default css`
       outline-offset: -1px;
     }
   }
+
+  ::slotted(sl-menu) {
+    max-width: var(--auto-size-available-width) !important;
+    max-height: var(--auto-size-available-height) !important;
+  }
 `;